+++ app/javascript/frontend/components/Compare/Models/Speed/index.vue
@@ -1,12 +1,7 @@
 <template>
   <div>
     <div class="row compare-row compare-section">
-      <div
-        :style="{
-          left: `${scrollLeft}px`,
-        }"
-        class="col-12 compare-row-label"
-      >
+      <div class="col-12 compare-row-label sticky-left">
         <div
           :class="{
             active: visible,
@@ -27,10 +22,7 @@
     <BCollapse id="speed" :visible="visible">
       <div class="row compare-row">
         <div
-          :style="{
-            left: `${scrollLeft}px`,
-          }"
-          class="col-12 compare-row-label text-right metrics-label"
+          class="col-12 compare-row-label text-right metrics-label sticky-left"
         >
           {{ $t('model.scmSpeed') }}
         </div>
@@ -46,10 +38,7 @@
       </div>
       <div class="row compare-row">
         <div
-          :style="{
-            left: `${scrollLeft}px`,
-          }"
-          class="col-12 compare-row-label text-right metrics-label"
+          class="col-12 compare-row-label text-right metrics-label sticky-left"
         >
           {{ $t('model.afterburnerSpeed') }}
         </div>
@@ -65,10 +54,7 @@
       </div>
       <div class="row compare-row">
         <div
-          :style="{
-            left: `${scrollLeft}px`,
-          }"
-          class="col-12 compare-row-label text-right metrics-label"
+          class="col-12 compare-row-label text-right metrics-label sticky-left"
         >
           {{ $t('model.groundSpeed') }}
         </div>
@@ -84,10 +70,7 @@
       </div>
       <div class="row compare-row">
         <div
-          :style="{
-            left: `${scrollLeft}px`,
-          }"
-          class="col-12 compare-row-label text-right metrics-label"
+          class="col-12 compare-row-label text-right metrics-label sticky-left"
         >
           {{ $t('model.afterburnerGroundSpeed') }}
         </div>
@@ -103,10 +86,7 @@
       </div>
       <div class="row compare-row">
         <div
-          :style="{
-            left: `${scrollLeft}px`,
-          }"
-          class="col-12 compare-row-label text-right metrics-label"
+          class="col-12 compare-row-label text-right metrics-label sticky-left"
         >
           {{ $t('model.pitchMax') }}
         </div>
@@ -122,10 +102,7 @@
       </div>
       <div class="row compare-row">
         <div
-          :style="{
-            left: `${scrollLeft}px`,
-          }"
-          class="col-12 compare-row-label text-right metrics-label"
+          class="col-12 compare-row-label text-right metrics-label sticky-left"
         >
           {{ $t('model.yawMax') }}
         </div>
@@ -141,10 +118,7 @@
       </div>
       <div class="row compare-row">
         <div
-          :style="{
-            left: `${scrollLeft}px`,
-          }"
-          class="col-12 compare-row-label text-right metrics-label"
+          class="col-12 compare-row-label text-right metrics-label sticky-left"
         >
           {{ $t('model.rollMax') }}
         </div>
@@ -160,10 +134,7 @@
       </div>
       <div class="row compare-row">
         <div
-          :style="{
-            left: `${scrollLeft}px`,
-          }"
-          class="col-12 compare-row-label text-right metrics-label"
+          class="col-12 compare-row-label text-right metrics-label sticky-left"
         >
           {{ $t('model.xaxisAcceleration') }}
         </div>
@@ -179,10 +150,7 @@
       </div>
       <div class="row compare-row">
         <div
-          :style="{
-            left: `${scrollLeft}px`,
-          }"
-          class="col-12 compare-row-label text-right metrics-label"
+          class="col-12 compare-row-label text-right metrics-label sticky-left"
         >
           {{ $t('model.yaxisAcceleration') }}
         </div>
@@ -198,10 +166,7 @@
       </div>
       <div class="row compare-row">
         <div
-          :style="{
-            left: `${scrollLeft}px`,
-          }"
-          class="col-12 compare-row-label text-right metrics-label"
+          class="col-12 compare-row-label text-right metrics-label sticky-left"
         >
           {{ $t('model.zaxisAcceleration') }}
         </div>
@@ -219,42 +184,32 @@
   </div>
 </template>
 
-<script>
+<script lang="ts">
+import Vue from 'vue'
+import { Component, Prop, Watch } from 'vue-property-decorator'
 import { BCollapse } from 'bootstrap-vue'
 
-export default {
+@Component<ModelsCompareSpeed>({
   components: {
     BCollapse,
   },
+})
+export default class ModelsCompareSpeed extends Vue {
+  @Prop({ required: true }) models!: Model[]
 
-  props: {
-    models: {
-      type: Array,
-      required: true,
-    },
+  visible: boolean = false
 
-    scrollLeft: {
-      type: Number,
-      required: true,
-    },
-  },
+  mounted() {
+    this.visible = this.models.length > 0
+  }
 
-  data() {
-    return {
-      visible: this.models.length > 0,
-    }
-  },
+  @Watch('models')
+  onModelsChange() {
+    this.visible = this.models.length > 0
+  }
 
-  watch: {
-    models() {
-      this.visible = this.models.length > 0
-    },
-  },
-
-  methods: {
-    toggle() {
-      this.visible = !this.visible
-    },
-  },
+  toggle() {
+    this.visible = !this.visible
+  }
 }
 </script>