+++ app/javascript/frontend/pages/Models/Compare/index.vue
@@ -1,8 +1,7 @@
 <template>
   <section
     :class="{
-      'scrolled-top': scrolledTop,
-      'scrolled-left': scrolledLeft,
+      'nav-slim': navSlim,
     }"
     class="container compare-models"
   >
@@ -61,12 +60,16 @@
                     <i class="fal fa-times" />
                   </div>
                 </div>
-                <div
-                  :style="{
-                    top: `${scrolledTopOffset - 81}px`,
-                  }"
-                  class="text-center compare-title"
-                >
+              </div>
+            </div>
+            <div class="row compare-row compare-row-headline sticky">
+              <div class="col-12 compare-row-label" />
+              <div
+                v-for="model in sortedModels"
+                :key="`${model.slug}-title`"
+                class="col-12 compare-row-item"
+              >
+                <div class="text-center compare-title">
                   <strong>{{ model.name }}</strong>
                 </div>
               </div>
@@ -81,26 +84,10 @@
               </div>
             </div>
             <div class="compare-wrapper">
-              <BaseRows
-                :models="sortedModels"
-                :scroll-left="scrolledLeftOffset"
-              />
-
-              <CrewRows
-                :models="sortedModels"
-                :scroll-left="scrolledLeftOffset"
-              />
-
-              <SpeedRows
-                :models="sortedModels"
-                :scroll-left="scrolledLeftOffset"
-              />
-
-              <CategoryRows
-                :models="sortedModels"
-                :scroll-left="scrolledLeftOffset"
-              />
-
+              <BaseRows :models="sortedModels" />
+              <CrewRows :models="sortedModels" />
+              <SpeedRows :models="sortedModels" />
+              <CategoryRows :models="sortedModels" />
               <Legend :models="sortedModels" />
             </div>
           </div>
@@ -110,7 +97,10 @@
   </section>
 </template>
 
-<script>
+<script lang="ts">
+import Vue from 'vue'
+import { Component, Watch } from 'vue-property-decorator'
+import { Getter } from 'vuex-class'
 import MetaInfo from 'frontend/mixins/MetaInfo'
 import FilterGroup from 'frontend/core/components/Form/FilterGroup'
 import Box from 'frontend/core/components/Box'
@@ -122,9 +112,7 @@ import SpeedRows from 'frontend/components/Compare/Models/Speed'
 import CategoryRows from 'frontend/components/Compare/Models/Categories'
 import Legend from 'frontend/components/Compare/Models/Legend'
 
-export default {
-  name: 'ModelsCompare',
-
+@Component<ModelsCompare>({
   components: {
     FilterGroup,
     Box,
@@ -136,159 +124,137 @@ export default {
     CategoryRows,
     Legend,
   },
-
   mixins: [MetaInfo],
+})
+export default class ModelsCompare extends Vue {
+  @Getter('navSlim', { namespace: 'app' }) navSlim: boolean
 
-  data() {
-    const query = JSON.parse(JSON.stringify(this.$route.query || {}))
+  newModel: Model | null = null
 
-    return {
-      scrolledTop: false,
-      scrolledLeft: false,
-      scrolledTopOffset: 0,
-      scrolledLeftOffset: 0,
-      newModel: null,
-      form: {
-        models: query.models || [],
-      },
-      models: [],
-    }
-  },
+  models: Model[] = []
 
-  computed: {
-    erkulUrl() {
-      return 'https://www.erkul.games/calculator'
-    },
+  form = {}
 
-    sortedModels() {
-      const models = JSON.parse(JSON.stringify(this.models))
+  get erkulUrl() {
+    return 'https://www.erkul.games/calculator'
+  }
 
-      return models.sort((a, b) => {
-        if (a.name < b.name) {
-          return -1
-        }
+  get sortedModels() {
+    const models = JSON.parse(JSON.stringify(this.models))
 
-        if (a.name > b.name) {
-          return 1
-        }
-
-        return 0
-      })
-    },
-
-    selectDisabled() {
-      return this.models.length > 7
-    },
-
-    disabledTooltip() {
-      if (this.selectDisabled) {
-        return this.$t('labels.compare.enough')
+    return models.sort((a, b) => {
+      if (a.name < b.name) {
+        return -1
       }
 
-      return null
-    },
+      if (a.name > b.name) {
+        return 1
+      }
 
-    crumbs() {
-      return [
-        {
-          to: {
-            name: 'models',
-          },
-          label: this.$t('nav.models.index'),
+      return 0
+    })
+  }
+
+  get selectDisabled() {
+    return this.models.length > 7
+  }
+
+  get disabledTooltip() {
+    if (this.selectDisabled) {
+      return this.$t('labels.compare.enough')
+    }
+
+    return null
+  }
+
+  get crumbs() {
+    return [
+      {
+        to: {
+          name: 'models',
         },
-      ]
-    },
-  },
-
-  watch: {
-    form: {
-      handler() {
-        this.update()
+        label: this.$t('nav.models.index'),
       },
-      deep: true,
-    },
-  },
+    ]
+  }
 
-  created() {
-    window.addEventListener('scroll', this.handleScroll)
-  },
-
-  destroyed() {
-    window.removeEventListener('scroll', this.handleScroll)
-  },
+  @Watch('form', { deep: true })
+  onFormChange() {
+    this.update()
+  }
 
   mounted() {
+    this.setupForm()
     this.form.models.forEach(async slug => {
       const model = await this.fetchModel(slug)
       this.models.push(model)
     })
-  },
+  }
 
-  methods: {
-    handleScroll() {
-      this.scrolledTop = document.documentElement.scrollTop > 261
-      this.scrolledLeft = document.documentElement.scrollLeft > 200
-      this.scrolledTopOffset = document.documentElement.scrollTop
-      this.scrolledLeftOffset = this.scrolledLeft
-        ? document.documentElement.scrollLeft
-        : 0
-    },
+  setupForm() {
+    const query = JSON.parse(JSON.stringify(this.$route.query || {}))
+    this.form = {
+      models: query.models || [],
+    }
+  }
 
-    update() {
-      this.$router.replace({
+  update() {
+    this.$router
+      .replace({
         name: this.$route.name,
         query: {
           models: this.form.models,
         },
       })
-    },
+      // eslint-disable-next-line @typescript-eslint/no-empty-function
+      .catch(() => {})
+  }
 
-    async add() {
-      if (this.newModel && !this.form.models.includes(this.newModel)) {
-        const model = await this.fetchModel(this.newModel)
-        this.models.push(model)
-        this.form.models.push(this.newModel)
-      }
-      this.newModel = null
-    },
+  async add() {
+    if (this.newModel && !this.form.models.includes(this.newModel)) {
+      const model = await this.fetchModel(this.newModel)
+      this.models.push(model)
+      this.form.models.push(this.newModel)
+    }
+    this.newModel = null
+  }
 
-    remove(model) {
-      if (this.form.models.includes(model.slug)) {
-        const index = this.form.models.indexOf(model.slug)
-        this.form.models.splice(index, 1)
-      }
+  remove(model) {
+    if (this.form.models.includes(model.slug)) {
+      const index = this.form.models.indexOf(model.slug)
+      this.form.models.splice(index, 1)
+    }
 
-      if (this.models.findIndex(item => item.slug === model.slug) >= 0) {
-        const index = this.models.findIndex(item => item.slug === model.slug)
-        this.models.splice(index, 1)
-      }
-    },
+    if (this.models.findIndex(item => item.slug === model.slug) >= 0) {
+      const index = this.models.findIndex(item => item.slug === model.slug)
+      this.models.splice(index, 1)
+    }
+  }
 
-    fetchModels({ page, search, missingValue }) {
-      const query = {
-        q: {},
-      }
+  fetchModels({ page, search, missingValue }) {
+    const query = {
+      q: {},
+    }
 
-      if (search) {
-        query.q.nameCont = search
-      } else if (missingValue) {
-        query.q.nameCont = missingValue
-      } else if (page) {
-        query.page = page
-      }
+    if (search) {
+      query.q.nameCont = search
+    } else if (missingValue) {
+      query.q.nameCont = missingValue
+    } else if (page) {
+      query.page = page
+    }
 
-      return this.$api.get('models', query)
-    },
+    return this.$api.get('models', query)
+  }
 
-    async fetchModel(slug) {
-      const response = await this.$api.get(`models/${slug}`)
+  async fetchModel(slug) {
+    const response = await this.$api.get(`models/${slug}`)
 
-      if (!response.error) {
-        return response.data
-      }
+    if (!response.error) {
+      return response.data
+    }
 
-      return null
-    },
-  },
+    return null
+  }
 }
 </script>