+++ app/javascript/frontend/plugins/LazyLoad/index.ts
@@ -3,4 +3,5 @@ import VueLazyload from 'vue-lazyload'
 
 Vue.use(VueLazyload, {
   observer: true,
+  loading: null,
 })