+++ db/schema.rb
@@ -17,7 +17,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
   enable_extension "pgcrypto"
   enable_extension "plpgsql"
 
-  create_table "admin_users", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "admin_users", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "username", limit: 255, default: "", null: false
     t.string "email", limit: 255, default: "", null: false
     t.string "encrypted_password", limit: 255, default: "", null: false
@@ -37,7 +37,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["username"], name: "index_admin_users_on_username", unique: true
   end
 
-  create_table "affiliations", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "affiliations", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "affiliationable_type"
     t.uuid "affiliationable_id"
     t.uuid "faction_id"
@@ -76,7 +76,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["visit_token"], name: "index_ahoy_visits_on_visit_token", unique: true
   end
 
-  create_table "albums", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "albums", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name", limit: 255
     t.string "slug", limit: 255
     t.boolean "enabled", default: false, null: false
@@ -84,7 +84,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.datetime "updated_at"
   end
 
-  create_table "celestial_objects", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "celestial_objects", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name"
     t.string "slug"
     t.uuid "starsystem_id"
@@ -111,7 +111,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["starsystem_id"], name: "index_celestial_objects_on_starsystem_id"
   end
 
-  create_table "commodities", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "commodities", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name"
     t.string "slug"
     t.datetime "created_at", null: false
@@ -134,7 +134,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["shop_commodity_id"], name: "index_commodity_prices_on_shop_commodity_id"
   end
 
-  create_table "components", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "components", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name", limit: 255
     t.string "size", limit: 255
     t.datetime "created_at"
@@ -151,7 +151,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["manufacturer_id"], name: "index_components_on_manufacturer_id"
   end
 
-  create_table "docks", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "docks", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.integer "dock_type"
     t.uuid "station_id"
     t.string "name"
@@ -168,7 +168,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["station_id"], name: "index_docks_on_station_id"
   end
 
-  create_table "equipment", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "equipment", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name"
     t.string "slug"
     t.string "store_image"
@@ -191,7 +191,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["manufacturer_id"], name: "index_equipment_on_manufacturer_id"
   end
 
-  create_table "factions", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "factions", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.integer "rsi_id"
     t.string "name"
     t.string "slug"
@@ -201,7 +201,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.datetime "updated_at", null: false
   end
 
-  create_table "fleet_memberships", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "fleet_memberships", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.uuid "fleet_id"
     t.uuid "user_id"
     t.integer "role"
@@ -215,7 +215,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.uuid "hangar_group_id"
   end
 
-  create_table "fleets", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "fleets", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "fid"
     t.string "slug"
     t.string "sid"
@@ -235,7 +235,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["fid"], name: "index_fleets_on_fid", unique: true
   end
 
-  create_table "habitations", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "habitations", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name"
     t.integer "habitation_type"
     t.uuid "station_id"
@@ -245,7 +245,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["station_id"], name: "index_habitations_on_station_id"
   end
 
-  create_table "hangar_groups", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "hangar_groups", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name"
     t.string "slug"
     t.string "color"
@@ -256,7 +256,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["user_id"], name: "index_hangar_groups_on_user_id"
   end
 
-  create_table "hardpoints", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "hardpoints", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.integer "quantity"
     t.uuid "model_id"
     t.uuid "component_id"
@@ -275,7 +275,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["model_id"], name: "index_hardpoints_on_model_id"
   end
 
-  create_table "images", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "images", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name", limit: 255
     t.uuid "gallery_id"
     t.string "gallery_type", limit: 255
@@ -290,7 +290,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["gallery_id"], name: "index_images_on_gallery_id"
   end
 
-  create_table "manufacturers", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "manufacturers", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name", limit: 255
     t.string "slug", limit: 255
     t.string "known_for", limit: 255
@@ -302,7 +302,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.string "code"
   end
 
-  create_table "model_additions", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "model_additions", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.uuid "model_id", null: false
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
@@ -321,14 +321,14 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["model_id"], name: "index_model_additions_on_model_id"
   end
 
-  create_table "model_loaners", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "model_loaners", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.uuid "model_id"
     t.uuid "loaner_model_id"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false
   end
 
-  create_table "model_modules", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "model_modules", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name"
     t.string "slug"
     t.text "description"
@@ -343,7 +343,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.decimal "pledge_price", precision: 15, scale: 2
   end
 
-  create_table "model_paints", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "model_paints", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name"
     t.uuid "model_id"
     t.string "slug"
@@ -371,7 +371,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.string "fleetchart_image"
   end
 
-  create_table "model_upgrades", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "model_upgrades", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name"
     t.string "slug"
     t.text "description"
@@ -383,7 +383,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.datetime "updated_at", null: false
   end
 
-  create_table "models", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "models", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name", limit: 255
     t.string "slug", limit: 255
     t.text "description"
@@ -468,14 +468,14 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["base_model_id"], name: "index_models_on_base_model_id"
   end
 
-  create_table "module_hardpoints", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "module_hardpoints", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.uuid "model_id"
     t.uuid "model_module_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
 
-  create_table "roadmap_items", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "roadmap_items", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.integer "rsi_id"
     t.integer "rsi_category_id"
     t.integer "rsi_release_id"
@@ -503,7 +503,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.datetime "updated_at", precision: 6, null: false
   end
 
-  create_table "shop_commodities", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "shop_commodities", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.uuid "shop_id"
     t.decimal "buy_price", precision: 15, scale: 2
     t.decimal "sell_price", precision: 15, scale: 2
@@ -529,7 +529,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["shop_id"], name: "index_shop_commodities_on_shop_id"
   end
 
-  create_table "shops", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "shops", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name"
     t.string "slug"
     t.datetime "created_at", null: false
@@ -546,7 +546,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["station_id"], name: "index_shops_on_station_id"
   end
 
-  create_table "star_citizen_updates", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "star_citizen_updates", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "url"
     t.string "title"
     t.string "news_type"
@@ -556,7 +556,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.datetime "updated_at", precision: 6, null: false
   end
 
-  create_table "starsystems", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "starsystems", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name"
     t.string "slug"
     t.datetime "created_at", null: false
@@ -581,7 +581,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.string "map_x"
   end
 
-  create_table "stations", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "stations", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "name"
     t.string "slug"
     t.uuid "planet_id"
@@ -603,7 +603,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["planet_id"], name: "index_stations_on_planet_id"
   end
 
-  create_table "task_forces", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "task_forces", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.uuid "hangar_group_id"
     t.uuid "vehicle_id"
     t.datetime "created_at", null: false
@@ -612,7 +612,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["vehicle_id"], name: "index_task_forces_on_vehicle_id"
   end
 
-  create_table "trade_routes", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "trade_routes", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.uuid "origin_id"
     t.uuid "origin_station_id"
     t.uuid "origin_celestial_object_id"
@@ -638,14 +638,14 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["origin_station_id"], name: "index_trade_routes_on_origin_station_id"
   end
 
-  create_table "upgrade_kits", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "upgrade_kits", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.uuid "model_id"
     t.uuid "model_upgrade_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
 
-  create_table "users", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "users", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "locale", limit: 255
     t.string "username", limit: 255, default: "", null: false
     t.string "email", limit: 255, default: "", null: false
@@ -684,21 +684,21 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["username"], name: "index_users_on_username", unique: true
   end
 
-  create_table "vehicle_modules", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "vehicle_modules", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.uuid "model_module_id"
     t.uuid "vehicle_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
 
-  create_table "vehicle_upgrades", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "vehicle_upgrades", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.uuid "model_upgrade_id"
     t.uuid "vehicle_id"
     t.datetime "created_at", null: false
     t.datetime "updated_at", null: false
   end
 
-  create_table "vehicles", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "vehicles", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.uuid "user_id"
     t.uuid "model_id"
     t.string "name", limit: 255
@@ -718,7 +718,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["user_id"], name: "index_vehicles_on_user_id"
   end
 
-  create_table "versions", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "versions", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "item_type", null: false
     t.uuid "item_id", null: false
     t.string "event", null: false
@@ -729,7 +729,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["item_type", "item_id"], name: "index_versions_on_item_type_and_item_id"
   end
 
-  create_table "videos", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "videos", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "url"
     t.integer "video_type"
     t.datetime "created_at", null: false
@@ -738,7 +738,7 @@ ActiveRecord::Schema.define(version: 2020_11_18_220722) do
     t.index ["model_id"], name: "index_videos_on_model_id"
   end
 
-  create_table "youtube_updates", id: :uuid, default: -> { "public.gen_random_uuid()" }, force: :cascade do |t|
+  create_table "youtube_updates", id: :uuid, default: -> { "gen_random_uuid()" }, force: :cascade do |t|
     t.string "video_id"
     t.datetime "created_at", precision: 6, null: false
     t.datetime "updated_at", precision: 6, null: false